+++ website/samples/samples_v3/custom_cms_view/CustomViewSampleApp.tsx
@@ -1,4 +1,4 @@
-import { buildCollection, CMSView, FireCMSApp, FireCMSAppConfig } from "@firecms/cloud";
+import { buildCollection, CMSView, FireCMSCloudApp, FireCMSAppConfig } from "@firecms/cloud";
 import { ExampleCMSView } from "./ExampleCMSView";
 
 const projectId = "YOUR_PROJECT_ID";
@@ -34,7 +34,7 @@ const appConfig: FireCMSAppConfig = {
 
 export default function App() {
 
-    return <FireCMSApp
+    return <FireCMSCloudApp
         projectId={projectId}
         appConfig={appConfig}
     />;