+++ CHANGELOG.md
@@ -1,9 +1,15 @@
+## [3.0.0-beta.5] - 2024-04-01
+
+- [BREAKING] The main component for FireCMS Cloud has been renamed from `FireCMSApp` to `FireCMSCloudApp`. Please update
+  your imports accordingly.
+- Fixes related to the CLI. You can now install the CLI globally with `npm install -g @firecms/cli`.
+
 ## [3.0.0-beta.4] - 2024-03-27
 
 - [BREAKING] The package name for FireCMS Cloud has changed from `firecms` to `@firecms/cloud`. This is done
   to avoid conflicts with the main FireCMS package. If you are using FireCMS Cloud, you will need to update your
   imports.
-- [BREAKING] If you are importing the tailwind configuration, you can now find the import at: 
+- [BREAKING] If you are importing the tailwind configuration, you can now find the import at:
   `import fireCMSConfig from "@firecms/ui/tailwind.config.js";`
 - [BREAKING] In that case, you also need to add `@tailwindcss/typography` to your dev dependencies.
 - [BREAKING] You need to update your `vite.config.js` and replace the package name in the federated configuration: